Removing cli module to move cli components closer to entrypoint  (#256)
* running model

* lint

* clean

* revert devtools

* fix comment

* lint
